Created component to display teams as cards on the discovery view
